Set up series nav in asyncComponent module
