+++ client/js/actions.js
@@ -1,2 +1,1 @@
 export const UPDATE_STICKY_NAV_HEIGHT = 'UPDATE_STICKY_NAV_HEIGHT';
-export const UPDATE_ASYNC_COMPONENTS = 'UPDATE_ASYNC_COMPONENTS';
