+++ server/views/pages/article.njk
@@ -9,10 +9,10 @@
 {% block body %}
 
 {% block seriesNavigation %}
-  {% if article.series[0] and article.series[0].commissionedLength %}
+  {#{% if article.series[0] and article.series[0].commissionedLength %}#}
     {% set asyncContentEndpoint = '/series-nav/' + article.series[0].url + '?current=' + article.url %}
     {% component 'async', { endpoint: asyncContentEndpoint, component: 'series-nav' } %}
-  {% endif %}
+  {#{% endif %}#}
 {% endblock %}
 {% componentV2 'page-description', {title: article.headline, meta: {type: 'date', value: article.datePublished}, lead: true}, {'a': true}, {icon: article.contentType | getIconForContentType} %}
